+++ primitive/update.py
@@ -30,6 +30,7 @@ from .monkey import Monkey
 from .oiltank import OilTank
 from .plane import Plane
 from .pyramid import Pyramid
+from .quadsphere import QuadSphere
 from .sphere import Sphere
 from .teapot import Teapot
 from .torus import Torus
@@ -61,6 +62,7 @@ def get_class(name):
 	elif name == "OilTank": return OilTank()
 	elif name == "Plane": return Plane()
 	elif name == "Pyramid": return Pyramid()
+	elif name == "QuadSphere": return QuadSphere()
 	elif name == "Sphere": return Sphere()
 	elif name == "Teapot": return Teapot()
 	elif name == "Torus": return Torus()
@@ -179,6 +181,8 @@ class PrimitiveData(PropertyGroup):
 	mode: StringProperty(update=primitive_update)
 	target: StringProperty(update=primitive_update)
 
+	# Profile ###############################################
+
 	profilo_mode: EnumProperty(name='Shape', default='Angle',
 		update = primitive_update,
 		items =[('Angle', 'Angle', ''),
@@ -197,6 +201,309 @@ class PrimitiveData(PropertyGroup):
 			('Manual', 'Manual', ''),
 			('Optimized', 'Optimized', ''),
 			('Adaptive', 'Adaptive', '')])
+
+	# BoltFactory #############################################
+
+	MAX_INPUT_NUMBER = 50
+
+	Bolt : BoolProperty(name = "Bolt", default = True, description = "Bolt")
+	change : BoolProperty(name = "Change",
+				default = False,
+				description = "change Bolt")
+
+	# Model Types
+	Model_Type_List = [('bf_Model_Bolt', 'BOLT', 'Bolt Model'),
+						('bf_Model_Nut', 'NUT', 'Nut Model')]
+	bf_Model_Type: EnumProperty(
+			attr='bf_Model_Type',
+			name='Model',
+			description='Choose the type off model you would like',
+			items=Model_Type_List, default='bf_Model_Bolt'
+			)
+	# Head Types
+	Model_Type_List = [('bf_Head_Hex', 'HEX', 'Hex Head'),
+						('bf_Head_12Pnt', '12 POINT', '12 Point Head'),
+						('bf_Head_Cap', 'CAP', 'Cap Head'),
+						('bf_Head_Dome', 'DOME', 'Dome Head'),
+						('bf_Head_Pan', 'PAN', 'Pan Head'),
+						('bf_Head_CounterSink', 'COUNTER SINK', 'Counter Sink Head')]
+	bf_Head_Type: EnumProperty(
+			attr='bf_Head_Type',
+			name='Head',
+			description='Choose the type off Head you would like',
+			items=Model_Type_List, default='bf_Head_Hex'
+			)
+	# Bit Types
+	Bit_Type_List = [('bf_Bit_None', 'NONE', 'No Bit Type'),
+					('bf_Bit_Allen', 'ALLEN', 'Allen Bit Type'),
+					('bf_Bit_Torx', 'TORX', 'Torx Bit Type'),
+					('bf_Bit_Philips', 'PHILLIPS', 'Phillips Bit Type')]
+	bf_Bit_Type: EnumProperty(
+			attr='bf_Bit_Type',
+			name='Bit Type',
+			description='Choose the type of bit to you would like',
+			items=Bit_Type_List, default='bf_Bit_None'
+			)
+	# Nut Types
+	Nut_Type_List = [('bf_Nut_Hex', 'HEX', 'Hex Nut'),
+					('bf_Nut_Lock', 'LOCK', 'Lock Nut'),
+					('bf_Nut_12Pnt', '12 POINT', '12 Point Nut')]
+	bf_Nut_Type: EnumProperty(
+			attr='bf_Nut_Type',
+			name='Nut Type',
+			description='Choose the type of nut you would like',
+			items=Nut_Type_List, default='bf_Nut_Hex'
+			)
+	# Shank Types
+	bf_Shank_Length: FloatProperty(
+			attr='bf_Shank_Length',
+			name='Shank Length', default=0,
+			min=0, soft_min=0, max=MAX_INPUT_NUMBER,
+			description='Length of the unthreaded shank',
+			unit='LENGTH',
+			)
+	bf_Shank_Dia: FloatProperty(
+			attr='bf_Shank_Dia',
+			name='Shank Dia', default=3,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Diameter of the shank',
+			unit='LENGTH',
+			)
+	bf_Phillips_Bit_Depth: FloatProperty(
+			attr='bf_Phillips_Bit_Depth',
+			name='Bit Depth', default=1.1431535482406616,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Depth of the Phillips Bit',
+			unit='LENGTH',
+			)
+	bf_Allen_Bit_Depth: FloatProperty(
+			attr='bf_Allen_Bit_Depth',
+			name='Bit Depth', default=1.5,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Depth of the Allen Bit',
+			unit='LENGTH',
+			)
+	bf_Allen_Bit_Flat_Distance: FloatProperty(
+			attr='bf_Allen_Bit_Flat_Distance',
+			name='Flat Dist', default=2.5,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Flat Distance of the Allen Bit',
+			unit='LENGTH',
+			)
+	# Torx Size Types
+	Torx_Size_Type_List = [('bf_Torx_T10', 'T10', 'T10'),
+					('bf_Torx_T20', 'T20', 'T20'),
+					('bf_Torx_T25', 'T25', 'T25'),
+					('bf_Torx_T30', 'T30', 'T30'),
+					('bf_Torx_T40', 'T40', 'T40'),
+					('bf_Torx_T50', 'T50', 'T50'),
+					('bf_Torx_T55', 'T55', 'T55'),
+					]
+
+	bf_Torx_Size_Type: EnumProperty(
+			attr='bf_Torx_Size_Type',
+			name='Torx Size',
+			description='Size of the Torx Bit',
+			items=Torx_Size_Type_List, default='bf_Torx_T20'
+			)
+	bf_Torx_Bit_Depth: FloatProperty(
+			attr='bf_Torx_Bit_Depth',
+			name='Bit Depth', default=1.5,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Depth of the Torx Bit',
+			unit='LENGTH',
+			)
+	bf_Hex_Head_Height: FloatProperty(
+			attr='bf_Hex_Head_Height',
+			name='Head Height', default=2,
+			min=0, soft_min=0, max=MAX_INPUT_NUMBER,
+			description='Height of the Hex Head',
+			unit='LENGTH',
+			)
+	bf_Hex_Head_Flat_Distance: FloatProperty(
+			attr='bf_Hex_Head_Flat_Distance',
+			name='Flat Dist', default=5.5,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Flat Distance of the Hex Head',
+			unit='LENGTH',
+			)
+	bf_12_Point_Head_Height: FloatProperty(
+			attr='bf_12_Point_Head_Height',
+			name='Head Height', default=3.0,
+			min=0, soft_min=0, max=MAX_INPUT_NUMBER,
+			description='Height of the 12 Point Head',
+			unit='LENGTH',
+			)
+	bf_12_Point_Head_Flat_Distance: FloatProperty(
+			attr='bf_12_Point_Head_Flat_Distance',
+			name='Flat Dist', default=3.0,
+			min=0.001, soft_min=0,    #limit to 0.001 to avoid calculation error
+			max=MAX_INPUT_NUMBER,
+			description='Flat Distance of the 12 Point Head',
+			unit='LENGTH',
+			)
+	bf_12_Point_Head_Flange_Dia: FloatProperty(
+			attr='bf_12_Point_Head_Flange_Dia',
+			name='12 Point Head Flange Dia', default=5.5,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Flange diameter of the 12 point Head',
+			unit='LENGTH',
+			)
+	bf_CounterSink_Head_Dia: FloatProperty(
+			attr='bf_CounterSink_Head_Dia',
+			name='Head Dia', default=6.300000190734863,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Diameter of the Counter Sink Head',
+			unit='LENGTH',
+			)
+	bf_Cap_Head_Height: FloatProperty(
+			attr='bf_Cap_Head_Height',
+			name='Head Height', default=3,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Height of the Cap Head',
+			unit='LENGTH',
+			)
+	bf_Cap_Head_Dia: FloatProperty(
+			attr='bf_Cap_Head_Dia',
+			name='Head Dia', default=5.5,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Diameter of the Cap Head',
+			unit='LENGTH',
+			)
+	bf_Dome_Head_Dia: FloatProperty(
+			attr='bf_Dome_Head_Dia',
+			name='Dome Head Dia', default=5.599999904632568,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Length of the unthreaded shank',
+			unit='LENGTH',
+			)
+	bf_Pan_Head_Dia: FloatProperty(
+			attr='bf_Pan_Head_Dia',
+			name='Pan Head Dia', default=5.599999904632568,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Diameter of the Pan Head',
+			unit='LENGTH',
+			)
+
+	bf_Philips_Bit_Dia: FloatProperty(
+			attr='bf_Philips_Bit_Dia',
+			name='Bit Dia', default=1.8199999332427979,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Diameter of the Philips Bit',
+			unit='LENGTH',
+			)
+
+	bf_Thread_Length: FloatProperty(
+			attr='bf_Thread_Length',
+			name='Thread Length', default=6,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Length of the Thread',
+			unit='LENGTH',
+			)
+
+	bf_Major_Dia: FloatProperty(
+			attr='bf_Major_Dia',
+			name='Major Dia', default=3,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Outside diameter of the Thread',
+			unit='LENGTH',
+			)
+
+	bf_Pitch: FloatProperty(
+			attr='bf_Pitch',
+			name='Pitch', default=0.3499999940395355,
+			min=0.1, soft_min=0.1,
+			max=7.0,
+			description='Pitch if the thread',
+			unit='LENGTH',
+			)
+	bf_Minor_Dia: FloatProperty(
+			attr='bf_Minor_Dia',
+			name='Minor Dia', default=2.6211137771606445,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Inside diameter of the Thread',
+			unit='LENGTH',
+			)
+	bf_Crest_Percent: IntProperty(
+			attr='bf_Crest_Percent',
+			name='Crest Percent', default=10,
+			min=1, soft_min=1,
+			max=90,
+			description='Percent of the pitch that makes up the Crest',
+			)
+	bf_Root_Percent: IntProperty(
+			attr='bf_Root_Percent',
+			name='Root Percent', default=10,
+			min=1, soft_min=1,
+			max=90,
+			description='Percent of the pitch that makes up the Root',
+			)
+	bf_Div_Count: IntProperty(
+			attr='bf_Div_Count',
+			name='Div count', default=36,
+			min=4, soft_min=4,
+			max=4096,
+			description='Div count determine circle resolution',
+			)
+	bf_Hex_Nut_Height: FloatProperty(
+			attr='bf_Hex_Nut_Height',
+			name='Hex Nut Height', default=2.4000000953674316,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Height of the Hex Nut',
+			unit='LENGTH',
+			)
+	bf_Hex_Nut_Flat_Distance: FloatProperty(
+			attr='bf_Hex_Nut_Flat_Distance',
+			name='Hex Nut Flat Dist', default=5.5,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Flat distance of the Hex Nut',
+			unit='LENGTH',
+			)
+	bf_12_Point_Nut_Height: FloatProperty(
+			attr='bf_12_Point_Nut_Height',
+			name='12 Point Nut Height', default=2.4000000953674316,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Height of the 12 Point Nut',
+			unit='LENGTH',
+			)
+
+	bf_12_Point_Nut_Flat_Distance: FloatProperty(
+			attr='bf_12_Point_Nut_Flat_Distance',
+			name='12 Point Nut Flat Dist', default=3.0,
+			min=0.001, soft_min=0,    #limit to 0.001 to avoid calculation error
+			max=MAX_INPUT_NUMBER,
+			description='Flat distance of the 12 point Nut',
+			unit='LENGTH',
+			)
+	bf_12_Point_Nut_Flange_Dia: FloatProperty(
+			attr='bf_12_Point_Nut_Flange_Dia',
+			name='12 Point Nut Flange Dia', default=5.5,
+			min=0, soft_min=0,
+			max=MAX_INPUT_NUMBER,
+			description='Flange diameter of the 12 point Nut',
+			unit='LENGTH',
+			)
+	
+	# End of Bolt factory ###################################################
 
 
 
@@ -224,6 +531,6 @@ def unregister_update():
 	del bpy.types.Mesh.primitivedata
 	del bpy.types.Curve.primitivedata
 	del bpy.types.Scene.primitive_setting
-	
+
 	bpy.utils.unregister_class(PrimitiveData)
 	bpy.utils.unregister_class(Primitive_Option)